Change: type of delay in every tick from int to double

+++ src/Coin/Coin.hpp
@@ -7,7 +7,7 @@ class Coin : public Item {
 	public:
 		Coin(double x, double y, double width, double height, int nilai);
 		int getValue(); //mendapatkan nilai koin
-		void tick(int delay);
+		void tick(double delay);
 	private:
 		int value; //nilai koin
 };
